dont run HelmetSEO jquery in test environment using process.env.NODE_ENV and move relatedItems from RelatedItems.test.js into Constants.js
